Extract some generic log logic to `AbstractMutableEventLog`
